pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ollama-proxy"
version = "0.1.0"
description = "FastAPI proxy that maps Affine-compatible OpenAI/Anthropic requests to local Ollama models."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "Affine Ollama Proxy Maintainers"}]
dependencies = [
    "fastapi>=0.111.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "uvicorn[standard]>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "coverage>=7.5",
    "pytest-cov>=5.0",
    "httpx>=0.27.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "-q"

[tool.coverage.run]
branch = true
source = ["ollama_proxy"]
omit = ["*/gateway.py"]

[tool.coverage.report]
fail_under = 90
show_missing = true
